Create useTime hook and replace it with the code in threshholdgraph
