+++ lib/hooks/index.ts
@@ -1,4 +1,5 @@
 import useClickedOutside from './useClickedOutside';
 import useFutureGraphData from './useFutureGraphData';
+import useTime from './useTime';
 
-export { useClickedOutside, useFutureGraphData };
+export { useClickedOutside, useFutureGraphData, useTime };
